+++ source/Tubeshade.Server/Services/Background/SchedulerService.cs
@@ -13,7 +13,6 @@ using Npgsql;
 using Tubeshade.Data;
 using Tubeshade.Data.Tasks;
 using Tubeshade.Server.Configuration;
-using Tubeshade.Server.Configuration.Startup;
 
 namespace Tubeshade.Server.Services.Background;
 
@@ -26,26 +25,26 @@ public sealed class SchedulerService : BackgroundService
     private readonly IClock _clock;
     private readonly IServiceProvider _serviceProvider;
     private readonly IOptionsMonitor<SchedulerOptions> _optionsMonitor;
-    private readonly DatabaseMigrationStartupFilter _migrationStartupFilter;
+    private readonly TaskListenerService _taskListenerService;
 
     public SchedulerService(
         ILogger<SchedulerService> logger,
         IClock clock,
         IServiceProvider serviceProvider,
         IOptionsMonitor<SchedulerOptions> optionsMonitor,
-        DatabaseMigrationStartupFilter migrationStartupFilter)
+        TaskListenerService taskListenerService)
     {
         _logger = logger;
         _clock = clock;
         _serviceProvider = serviceProvider;
         _optionsMonitor = optionsMonitor;
-        _migrationStartupFilter = migrationStartupFilter;
+        _taskListenerService = taskListenerService;
     }
 
     /// <inheritdoc />
     protected override async Task ExecuteAsync(CancellationToken stoppingToken)
     {
-        await _migrationStartupFilter.MigrationTask;
+        await _taskListenerService.IsListeningTask.WaitAsync(stoppingToken);
 
         var timerPeriod = _optionsMonitor.CurrentValue.GetPeriodTimeSpan();
         using var timer = new PeriodicTimer(timerPeriod);